Module to read and write RFID tags using MFRC522 IC
The MFRC522 module in MicroPython is a class that allows users to read and write RFID tags using the MFRC522 IC. The MFRC522 is a low-cost, low-voltage, and small-sized integrated read and write card chip that operates at 13.56 MHz. It can communicate with a host machine via serial, SPI, or I2C mode.