The LCD API module in MicroPython
The LCD API module in MicroPython is a tool that allows users to communicate with character LCDs that are compatible with HD44780. The module knows which commands to send to the LCD, but it doesn't know how to get them to the LCD. A derived class is required to implement the hal_xxx functions.